bot/requirements.txt
python-telegram-bot==22.0
matplotlib==3.10.3
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
Telethon==1.40.0
statsmodels==0.14.4
transformers==4.51.3
torch==2.7.0
peft==0.15.2
python-dotenv==1.1.0 